I use these scripts to run the Hodgkin-Huxley model.
The HH model is written in cython. This is why it needs to be "cythonized" first. For this, run setup.py. This will cythonize HH_cython.pyx. This script needs HH_helper.py to run. HH_track keeps track of the simulated model results if used in an evolutionary algorithm.